datalist: fix issue with the browser "back" button
It seems that when autocomplete is on, we can hit enter with an
empty text field, and we will face a 404 in the request info page.
I tested this update, and it seems to solve the issue. However,
I was not able to reproduce the exact error that we've seen in production.

I also added the "select" css class to the dropdown to show the arrow
in the input field.

Signed-off-by: Jorge Ramírez <jorgeramirez1990@gmail.com>