templates: Enable map clustering by default
The app behaves  way too laggy in highly populated areas
therefore it might be worth considering a smarter way to
decide whether or not to use map clustering.

e.g. marks.length > threshold ? enabled : disabled;

Until then, it could be beneficial to the user to have
its default to "enabled".